requirements.txt
altair==4.2.2
apturl==0.5.2
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
cassandra-pylib==0.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.5
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.12.4
ptyprocess==0.7.0
pyarrow==12.0.0
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.1b0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.3.5
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.0
streamlit==1.22.0
systemd-python==234
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.20.0
virtualenv==20.15.1
wadllib==1.3.6
watchdog==3.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
